add readme, tests matmul

+++ include/mat_utils.h
@@ -1,9 +1,9 @@
 #include "stdio.h"
 #include "stdlib.h"
 
-float *generate_random_matrix(int rows, int cols);
+double *generate_random_matrix(int rows, int cols);
 
-void print_matrix(float *matrix, int rows, int cols);
+void print_matrix(double *matrix, int rows, int cols);
 
-void matrix_multiply(float *matrix1, float *matrix2, float *result, int m, int n, int k);
+void matrix_multiply(double *matrix1, double *matrix2, double *result, int m, int n, int k);
